feat: 매니저 등록 시 로깅

+++ src/main/java/org/example/expert/domain/todo/dto/response/TodoSearchResponse.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class TodoSearchResponse {
     private final String title;
-    private final int managerCount;
-    private final int commentCount;
+    private final Long managerCount;
+    private final Long commentCount;
 }
